rwinslow - Implemented CloudTrail declaration w/ and w/o explicit bucket name, only w/o tested
The cloud_trail declare command now creates a cloud_trail and, optionally, its underlying S3
bucket using the aws CLI's cloudtrail create-subscription command. The only scenario written
is for the auto-create-bucket option (as opposed to explicitly specifying a bucket name), as
this is expected to be the common case and obtaining a real output example for the explicit
bucket name case required bucket policy-setting work that was too hard to get done during
the hack-a-thon.

Also did some cleanup with the S3 bucket service to conform to error and output handling
conventions generally.